requirements.txt
flask-limiter
mysql-connector-python
python-dotenv
pyjwt
bcrypt
flask-cors
gunicorn
Flask
together
sentence-transformers
scikit-learn
numpy
python-dateutil
regex
pandas
huggingface_hub
transformers
torch
faiss-cpu
rank_bm25
langchain
langchain_experimental
langchain_together